requirement.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
jsmin==2.2.2
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
